moved: setters out of the Config class
