+++ Makefile
@@ -28,6 +28,7 @@ SRCS   	= \
         $(SRCS_DIR)/ConfigParser.cpp \
         $(SRCS_DIR)/ConfigDispatcher.cpp \
         $(SRCS_DIR)/Utils.cpp \
+        $(SRCS_DIR)/ConfigSetters.cpp \
 		$(SRCS_DIR)/main_config.cpp
 TESTS	= \
