+++ src/Config.cpp
@@ -16,6 +16,7 @@
 #include <sys/socket.h>
 #include <vector>
 #include <arpa/inet.h>
+#include "ConfigSetters.hpp"
 
 // full initialization of the config in one constructor
 // 
@@ -80,12 +81,13 @@ void	Config::set_servers(std::map <int, std::map <std::string, std::vector <std:
 
 			configure_server_names(it->second, new_server, new_unique_values);
 			configure_port(it->second, new_server, new_unique_values);
-			configure_access_log(it->second, new_server);
 			configure_host(it->second, new_server, new_unique_values);
-			configure_client_max_body_size(it->second, new_server);
-			configure_autoindex(it->second, new_server);
-			configure_root(it->second, new_server);
-			configure_index(it->second, new_server);
+			Setters::configure_access_log(it->second, new_server);
+			Setters::configure_client_max_body_size(it->second, new_server);
+			Setters::configure_autoindex(it->second, new_server);
+			Setters::configure_root(it->second, new_server);
+			Setters::configure_index(it->second, new_server);
+			
 			configure_locations(it->second, new_server);
 
 			_servers.push_back(new_server);
@@ -107,19 +109,7 @@ void	Config::set_servers(std::map <int, std::map <std::string, std::vector <std:
 //
 // else:
 //			->	return empty string
-std::string	Config::extract_location_name(const std::string& current_map_key)
-{
-	std::string location_key_prefix = "location";
-    size_t found = current_map_key.find(location_key_prefix);
 
-    if (found != std::string::npos) 
-	{
-        size_t colon = current_map_key.find(":", found);
-        return current_map_key.substr(found + location_key_prefix.size() + 1, colon - (found + location_key_prefix.size()) - 1);
-    }
-
-    return ""; 
-}
 
 // initializes the location before setting the values
 //
@@ -166,7 +156,7 @@ void	Config::configure_locations(const _map& server, Server*& new_server)
 	for (_map::const_iterator it = server.begin(); it != server.end(); it++)
 	{
 
-		std::string name = extract_location_name(it->first);
+		std::string name = Utils::extract_location_name(it->first);
 
 		if (name.empty() == true)
 		{
@@ -322,104 +312,7 @@ void	Config::configure_port(_map& server, Server*& new_server, std::vector <std:
 	new_unique_values.push_back(port);
 }
 
-// validates access log from config
-//
-// 	if:		the file does not exist (meaning we can just create it)
-// 	or:		it exists and we have write access to it
-//			->	config valid, return the filename
-//
-// 	else:
-//			->	path invalid, log error & return default 
-void	Config::configure_access_log(_map& server, Server*& new_server)
-{
-	std::string access_log;
 
-	if (server.find("access_log") != server.end() && server["access_log"].empty() == false)
-	{
-		access_log = server["access_log"][0];
-
-		if (Utils::file_exists(access_log) == false or Utils::write_access(access_log) == true)
-		{
-			new_server->set_access_log(access_log);
-			return ;
-		}
-	}
-	Log::log("error: access log '" + access_log + "' could not be opened, falling back to 'access.log'\n", STD_ERR | ERROR_FILE);
-
-	new_server->set_access_log(ACCESS_LOG_DEFAULT);
-}
-
-void	Config::configure_index(_map& server, Server*& new_server)
-{
-	std::string index;
-
-	if (server.find("index") != server.end() && server["index"].empty() == false)
-	{
-		index = server["index"][0];
-
-		if (Utils::file_exists(index) == false or Utils::write_access(index) == true)
-		{
-			new_server->set_index_path(index);
-			return ;
-		}
-	}
-	Log::log("error: access log '" + index + "' could not be opened, falling back to 'access.log'\n", STD_ERR | ERROR_FILE);
-	new_server->set_index_path("index.html");
-}
-
-void	Config::configure_autoindex(_map& server, Server*& new_server)
-{
-	if (server.find("autoindex") != server.end() && server["autoindex"].empty() == false)
-	{
-		if (server["autoindex"][0] == "enabled")
-		{
-			new_server->set_auto_index(true);
-		}
-	}
-}
-
-void	Config::configure_root(_map& server, Server*& new_server)
-{
-	if (server.find("root") != server.end() && server["root"].empty() == false)
-	{
-		new_server->set_root(server["root"][0]);
-	}
-}
-
-// validates client max body size from config
-//
-// if:	the value is missing or empty
-//		->	log error & fall back to default
-//
-// else if: the value is too high
-//		->	log error & cap it to 10M
-//
-// else if: the value is invalid
-//		->	log error & fall back to default
-void	Config::configure_client_max_body_size(_map& server, Server*& new_server)
-{
-	if (server.find("client_max_body_size") == server.end() or server["client_max_body_size"].empty() == true)
-	{
-		Log::log("no client max body size config in server '" + new_server->get_server_name()[0] + "', falling back to default (1M)\n", STD_ERR | ERROR_FILE);
-		new_server->set_client_max_body_size(CLIENT_MAX_BODY_SIZE_DEFAULT);
-		return ;
-	}
-
-	std::string client_max_body_size = server["client_max_body_size"][0];
-	int	size = Utils::parse_client_max_body_size(client_max_body_size);
-
-	if (size > CLIENT_MAX_BODY_SIZE_MAX)
-	{
-		Log::log("error: " + client_max_body_size + ": client max body size too high, capping to 10M\n", STD_ERR | ERROR_FILE);
-		size = CLIENT_MAX_BODY_SIZE_MAX;
-	}
-	else if (size == -1)
-	{
-		Log::log("error: client max body size '" + client_max_body_size + "' is not valid, falling back to default (1M)\n", STD_ERR | ERROR_FILE);
-		size = CLIENT_MAX_BODY_SIZE_DEFAULT;
-	}
-	new_server->set_client_max_body_size(size);
-}
 
 std::vector <Server *>	Config::get_servers() const
 {
@@ -438,102 +331,17 @@ std::string	Config::get_error_page(const int key)
 	}
 }
 
-void	set_root(const std::vector <std::string>& root, LocationInfo*& new_location)
-{
-	if (root.empty() == false)
-	{
-		new_location->set_root(root[0]);
-	}
-}
-
-void	set_directory_listing(const std::vector <std::string>& directory_listing, LocationInfo*& new_location)
-{
-	if (directory_listing.empty() == false)
-	{
-		if (directory_listing[0] == "enabled")
-		{
-			new_location->set_directory_listing(true);
-		}
-	}
-}
-
-void	set_allowed_methods(const std::vector <std::string>& allowed_methods, LocationInfo*& new_location)
-{
-	if (allowed_methods.empty() == false)
-	{
-		new_location->set_allowed_methods(allowed_methods);
-	}
-	else 
-	{
-		std::vector <std::string> foo;
-		foo.push_back("none");
-		new_location->set_allowed_methods(allowed_methods);
-	}
-}
-
-void	set_return(const std::vector <std::string>& rtrn, LocationInfo*& new_location)
-{
-	if (rtrn.empty() == false)
-	{
-		new_location->set_return(rtrn[0]);
-	}
-}
-
-void	set_alias(const std::vector <std::string>& alias, LocationInfo*& new_location)
-{
-	if (alias.empty() == false)
-	{
-		new_location->set_alias(alias);
-	}
-}
-
-void	set_cgi_path(const std::vector <std::string>& cgi_path, LocationInfo*& new_location)
-{
-	if (cgi_path.empty() == false)
-	{
-		new_location->set_cgi_path(cgi_path[0]);
-	}
-}
-
-
-void	set_cgi_extension(const std::vector <std::string>& extension, LocationInfo*& new_location)
-{
-	if (extension.empty() == false)
-	{
-		new_location->set_cgi_extensions(extension);
-	}
-}
-
-void	set_autoindex(const std::vector <std::string>& autoindex, LocationInfo*& new_location)
-{
-	if (autoindex.empty() == false)
-	{
-		if (autoindex[0] == "enabled")
-		{
-			new_location->set_autoindex(true);
-		}
-	}
-}
-
-void	set_index(const std::vector <std::string>& index, LocationInfo*& new_location)
-{
-	if (index.empty() == false)
-	{
-		new_location->set_index_path(index[0]);
-	}
-}
-
 void	Config::initialize_location_setters()
 {
-	_location_setters["root"] = &set_root;
-	_location_setters["directory_listing"] = set_directory_listing;
-	_location_setters["allowed_methods"] = set_allowed_methods;
-	_location_setters["return"] = set_return;
-	_location_setters["alias"] = set_alias;
-	_location_setters["handler"] = set_cgi_path;
-	_location_setters["extension"] = set_cgi_extension;
-	_location_setters["autoindex"] = set_autoindex;
-	_location_setters["index"] = set_index;
+	_location_setters["root"] = &Setters::set_root;
+	_location_setters["directory_listing"] = &Setters::set_directory_listing;
+	_location_setters["allowed_methods"] = &Setters::set_allowed_methods;
+	_location_setters["return"] = &Setters::set_return;
+	_location_setters["alias"] = &Setters::set_alias;
+	_location_setters["handler"] = &Setters::set_cgi_path;
+	_location_setters["extension"] = &Setters::set_cgi_extension;
+	_location_setters["autoindex"] = &Setters::set_autoindex;
+	_location_setters["index"] = &Setters::set_index;
 }
 
 Config::~Config()
